Updated handleSubmit function in login-view to post login credentials to the server using axios. Then updated main-view to update users state with login data and get movies.
If successfully authenticated, login data (user and token) are passed to the onLoggedIn method in main-view. This then gets the movies using the token for authentication. Local storage is set with the login data and used by the componentDidMount method to ensure the user remains logged in, even when the page is refreshed.