Created a separate navbar component with a functional logout button.
Because the navbar is the best place to hold the logout button, it has been moved from the index.jsx file to a separate file within the component folder. The button has been added and this is styled in a dedicated css file for the navbar. The logout function needs to update the state of the main-view component, so it is included within the MainView class component and passed as props to the navbar component, which invokes the function onclick. Button text for all components has been updated to ensure consistency across the add (capital letters removed).