Amended spacing of code next to return statements in main-view. Amended css for login and registration views to move content down the page. Moved button from registration view and added registration button to login view.
The code was not working with the columns on a separate line to the return statement. As we have put a redirect on the registration view to go to the login page after registering we no longer need a separate button. We do need a button on the login page to link to the registration page for new users.