Updated profile view to improve layout, add proptypes, fix deregister function (details in description)
The layout has been changed to a form to look consistent with other views. The data fetching has been moved inside a useEffect hook with a condition so that it only happens on the initial render. A condition has been added so that a statement returns if no movies have been added to favourites. Password is now taken from local storage rather than the database in order to display the unhashed version that the user will recognise. The login view has been updated to store this when logging in (it is immediately removed if the user is not recognised) and the logout function removes it.