Corrected update function that did not work previously when the username itself was updated and updated form to pre-populate user data from logindata stored in local storage.
Login function updated to store email and birthday as well as password and logout function updated to set values to null on logout. Note some javascript was required within the update component to convert the date to the correct format for populating the initial value of the form. Note also, password values have been changed from password type to text to assist with testing. They will be changed back in the final version of the app.