Updated onLoggedIn function and logOut function.
When logging in we now update the state with the full user object (not just username) so the data can be reused. LogOut has been updated to dispatch the SET_USER action and update the user in the store to null, rather than using setState within the component. The local storage items that may not be needed have been commented out until the new code is tested. MovieCard is no longer needed so has been removed.