A new function getUser has been created which is called within componentDidMount when the page refreshes.
This is used to reset the user state on a page refresh, in the same way that getMovies is called in both the log in and componentDidMount functions. Note that local storage is still used for the username and password to avoid the user logging in again on a page refresh.